// frontend/src/types/bookingResponseType.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type BookingResponseType = typeof BookingResponseType[keyof typeof BookingResponseType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BookingResponseType = {
  SELL_ADVICE: 'SELL_ADVICE',
  BUY_ADVICE: 'BUY_ADVICE',
  LEGAL_ADVICE: 'LEGAL_ADVICE',
  OTHER: 'OTHER',
} as const;