// frontend/src/types/bookingResponseStatus.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type BookingResponseStatus = typeof BookingResponseStatus[keyof typeof BookingResponseStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BookingResponseStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED',
} as const;